refactor(core): reuse connector reading function in cli
